Update Home and UserProfile component. The search button does not currently work and requires that other components are refactored in order to render the search results. Clicking on links in the Header and Sidebar will not break the app.
