+++ src/components/Home.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import Typography from "@material-ui/core/Typography";
 import Stats from "./Stats";
 
 const recentBlogs = "Your Friends' Popular/Recent blogs";
